user in celery worker

+++ django/celery/worker.sh
@@ -1,3 +1,7 @@
 #! /bin/sh
-rm proj/celery.log
-celery -A proj worker -l INFO
+if [ -e proj/celery.log ]
+then
+    rm proj/celery.log
+fi
+
+exec celery -A proj worker -l INFO
